EDA_DataDescription
In this file we have briefly described about our data.
We have made a boxplot over a time -frequency graph to get an overview of the data
We have also plotted the ACF and Decomp graph to check the trend and seasonality. We can see that there is seasonality and additive trend present in our dataset.
We have also plotted the time series graph, where we can see a clear seasonality in the values of this dataset.